refactor(svelte): extract controller pure canonical helpers
Move key/scope/domain/interval canonicalize and equality helpers out of
createPlotInteraction into controller-canonical.ts so pure logic is not
mixed with the Svelte controller factory. Public createPlotInteraction
export path unchanged.
